Update tabManager.js
Refactored the tab checking functionality. There is now a function which checks if a tab is labeled as distracting. There are now event listeners which check if the URL in the tab changes to a distracting one. There is still the issue of the locker activating immediately when I mark a tab as distracting